// src/components/MapPartComponents/data.d.ts
interface tabsObj{
    title:string,
    name:string,
    content:string,
    isActive:string
}
// ...
export const MapStateInit={
    codeMapData:[],             //码点地图数据
    gridMapData:{},             //格点地图数据
    portsListData:[],           //工位点列表数据
    drawMapParamsArr:[],
    pointLocateFlag:false,      //点位定位
    pathPointList:[],           //路径点位列表
    createPointFlag:false,     //创建点位列表
    createPathFlag:false,       //创建线路列表
    setPathInputFlag:false,
    reqLoadFlag:false,
    drawMapParams:{ 
        startPosX:100,            //起点X
        startPosY:100,            //起点Y
        startWide:40,            //起点码宽
        space:80,
        canvasWidth:0,          //画布宽
        canvasHeight:0,         //画布高
        gridPointWidth:0,       //格点宽度
        gridPointHeight:0,      //格点高度
        scaleProp:1,            //缩放比例
        mapWidth: 0,
        mapHeight: 0,
        colorDefine:[
            {idx:"0",color:"#3abae9",name:'idle',colorOut:"rgba(58,186,233,0.3)",pattern:"blank",tips:'空闲'},
            {idx:"1",color:"#e96b3a",name:'busy',colorOut:"rgba(233,107,58,0.3)",pattern:"blank",tips:'忙碌'},       
            {idx:"2",color:"#3ae9ad",name:'charging',colorOut:"rgba(58,233,173,0.3)",pattern:"blank",tips:'充电'},
            {idx:"3",color:"#DC191D",name:'malfunction',colorOut:"rgba(220,25,29,0.3)",pattern:"blank",tips:'故障'},
            {idx:"4",color:"#E0C718",name:'anticollision',colorOut:"rgba(224,199,24,0.3)",pattern:"blank",tips:'防碰撞'},
        ]
    },
    drawMapParamsCache:{
        startPosX: 0, //开始绘制x位置
        startPosY: 0, //开始绘制y位置
        space: 0,
        startWide: 0,   
        canvasWidth: 0,
        canvasHeight: 0,
        gridPointWidth: 0,
        gridPointHeight: 0,
        scaleProp: 0,
        mapWidth:0,
        mapHeight:0
    },
    currentFloor:0,      //当前楼层
    curFloorPorts:[],    //当前楼层工位
    curFloorGrids:[],    //当前楼层格点
    curPortsListArr:[],
    curLocPoint:null,
    regionsPointArr:<Array<any>>[],           //区域对象
    regionsPortArr:[],
    editableTabs:<Array<tabsObj>>[],
    gridDataArr:[],             //格点数据数组
    activeTab:'0',             //活跃标签
    loading:true,
    formLabelWidth: '120px',
    maxPortId:0,
    portShowType:'',
    actionList:[],
    pathList:{},            //线路列表
    bufferAreaList:[],      //缓冲区列表
    agvList:{},             //Agv列表
    agvTimeOut:0,           //小车定时器
    currentAgv:'',          //当前Agv
    currentAgvObj:'',       //当前追踪Agv对象
    monitorAgvList:[],      //模拟Agv列表
    allDDL:[],              //ddl列表
    allCodePortType:[],     //码点对应的站点类型  
    trackFlag:false,
    rightCollapse:false,      //侧边栏显隐
    allAgvList:'{}',
    firstMapInit:true,
    listStatus:[]
}


export interface StateType {
    dmCodeList?:Array<object>,
    gridPointList?:Array<object>,
    portTypeList?:Array<object>,
    ddlList?:Array<object>,
    pathList?:Array<object>,
    nowAgvState?: Array<object>,
    socketObj?:Object,
    exceptionList?: Array<object>,
    socketStatus?: String,
}